requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-doc==0.0.4
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
dataclasses-json==0.6.3
debugpy==1.8.0
decorator==5.1.1
distro==1.8.0
executing==2.0.1
fastapi==0.128.0
filelock==3.13.1
Flask==2.3.3
frozenlist==1.4.0
fsspec==2023.12.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.31.0
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.11.0
iniconfig==2.1.0
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.5.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
langchain==0.0.344
langchain-core==0.0.8
langsmith==0.0.67
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.6.1
numpy==2.2.6
oauthlib==3.3.1
openai==1.3.6
opencv-python==4.12.0.88
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
pluggy==1.6.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.8.1b0
Pygments==2.17.2
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
pyparsing==3.2.5
pypdf==6.6.0
pytest==8.4.2
python-dateutil==2.8.2
python-dotenv==1.1.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.0
rpds-py==0.13.2
rsa==4.9.1
safetensors==0.4.1
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.9.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.50.0
streamlit==1.29.0
sympy==1.14.0
tenacity==8.2.3
threadpoolctl==3.6.0
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.9.1
torchvision==0.24.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.35.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.40.0
validators==0.22.0
wcwidth==0.2.12
Werkzeug==3.1.3
widgetsnbextension==4.0.9
yarl==1.9.3
zipp==3.17.0
